Fix conv mapping of scalars
OpenMP 4.5 maps Scalar Variables as firstprivate, so the convergence
checking was broken here because the convergence checking target
region was assuming the conv variable was not firstprivate, but
because no data-sharing clauses said otherwise the spec assumed
firstprivate. This change makes no difference on Cray CCE, but
is OpenMP 4.5 conformant for clang-ykt.